Fix drag and drop of blocks
