+++ ExPlast/sitebuilder/frontend/modules/api.js
@@ -1,0 +1,9 @@
+export async function api(method, path, body) {
+  const r = await fetch(path, {
+    method,
+    headers: { 'Content-Type': 'application/json' },
+    body: body ? JSON.stringify(body) : undefined,
+  });
+  if (!r.ok) throw await r.text();
+  return r.json();
+}
