+++ ExPlast/sitebuilder/frontend/modules/constants.js
@@ -1,0 +1,2 @@
+export const AUTO_SAVE_KEY = 'builderAutosave';
+export const AUTO_SAVE_INTERVAL = 10000;
